Use a longer autokey period.
NOTE: this is a breaking change, texts encrypted with 1.0.3 or before
will not be able to be decrypted with 1.0.4!

Previously, only the last character in the key was contributing to
subsequent rotates. Now, the full key length should be used for texts
longer than the key.  This should make the algorithm more secure.

Note that the algorithm is NOT secure in the modern sense of the word.
If an attacker knows a statistical characteristic of the language of the
plain text (such as the n-gram frequencies of a human language or the
possibility set of a computer-encoded string) they will likely recover
the original key and decipher any future messages.