chore: move coveralls from package.json to build
Coveralls is only really necessary for the master build. So install that
specifically once the master is ready to do the coverage upload.

Also move the coverage run to the script section so tests don't have to
be run twice. Yeah, it'll put coverage reports on the minions but that
shouldn't be such a big deal.